requirements.txt
django
djangorestframework
flake8
autopep8
django-filter
markdown
djoser
drf-yasg 